new combins, fixed more bugs

+++ test/test_c2/test_c2.c++
@@ -51,10 +51,10 @@ int main( int argn, char* argv[] )
 
     FermionHamil Hop(ints);
     //Lets test hartree-fock
-    uint32_t u1 = 37793167;
-    uint32_t d1 = 37793167;
-    u1 = (1 << Nups)-1;
-    d1 = (1 << Ndos)-1;
+    uint64_t u1 = 37793167;
+    uint64_t d1 = 37793167;
+    u1 = (1LL << Nups)-1LL;
+    d1 = (1LL << Ndos)-1LL;
     uint64_t st = (d1 << Norbs) + u1;
     slater_det hello =  slater_det( st, Norbs, Nups, Ndos ) ;
     double nE =  Hop.GetHmatel(hello,hello);
